Seperate entrypoint for tempest plugins for python-aodh-tests
Previously python-aodh package installed the entry point for the in-tree
Tempest plugin which cause failure while running tempest list-plugins
because Tempest found the entry point but does not have Tempest plugin
code which is subpackaged in python-aodh-tests.
Separate aodh_tests egg-info is created with the plugin entry point
so that Tempest includes the plugin only when subpackage is installed.

Change-Id: I4e89bcafb3daed363b5c2eac8690b035794e9c10