Move aodh-dbsync to openstack-aodh-common
aodh-dbsync should be part of the -common package and not tied to the
api service which could be running on a different host entirely. Moving
it to -common also brings it inline with other OpenStack packages who
have their db management commands in their -common subpackages.

See https://bugs.launchpad.net/tripleo/+bug/1656066

Change-Id: Ib8f9ad166f9bbed17584b7d8f207b73439def53d